+++ tests/unit/flows/normalized/flow9.yml
@@ -1,0 +1,4 @@
+jtype: Flow
+executors:
+  - name: E1
+    uses: ${{ CONTEXT.E1_USES }}
